backend/requirements-sqlite.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.27
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic>=1.13.0
email-validator>=2.0.0
